Fixed bugs
- Fixed a bug where Tbl format is not properly parsing offset larger than ushort maximum
  - This also introduces a breaking change that makes normal Tbl file from fb not work if there's a subfolder flag, will properly implement the switch in the future
- Fixed a bug where UnpackFhm path endpoint is not working at all, and made two distinct commands to handle the UnpackFhm process
- Fixed a bug where FhmPacker wrongly parses the whole directory instead of top directory only when parsing subfolder directory, thus causing large file size
- Fixed a bug where subdirectory contents will be written twice for Fhm packing operation
- Fixed a bug where duplicate items in Fhm is not properly referenced and causes offset issues
-

+++ src/Core/Infrastructure/Formats/TblFormat/TblBinarySerializer.cs
@@ -12,7 +12,7 @@ public class TblBinarySerializer : IFormatBinarySerializer<Tbl>
     public Task<Tbl> DeserializeAsync(Stream data, CancellationToken cancellationToken)
     {
         var kaitaiStream = new KaitaiStream(data);
-        var deserializedObject = new Tbl((ushort)data.Length, kaitaiStream);
+        var deserializedObject = new Tbl((uint)data.Length, kaitaiStream);
         return Task.FromResult(deserializedObject);
     }
